refactor(delegation): optimize pending records
BREAKING: the pending records are now stored in big endian format of the
height (as opposed to the previous hex-encoded values). this allows us
to easily iterate over all of the records in a height-ascending manner
and determine the ones that should have mature at or equal to the
current block height.